cpu_runtime/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
structlog>=24.1.0
prometheus-client>=0.20.0
llama-cpp-python>=0.2.90
httpx>=0.27.0
